Send LabelsRotation as a real so fractional tick angles survive
Q's LabelsRotation is now a double, so use as.numeric instead of as.integer -
a fractional categories.tick.angle now survives the round trip (and, with the
companion q change, all the way to the exported axis, since PowerPoint stores
rotation in 60,000ths of a degree).

+++ R/cchart.R
@@ -861,7 +861,7 @@ getPPTSettings <- function(chart.type, args, data)
 
         # Also skip the horizontal/Automatic default angle (0), which needs no rotation.
         if (q.can.parse.date.axis && isTRUE(args$categories.tick.angle != 0))
-            res$PrimaryAxis$LabelsRotation <- as.integer(args$categories.tick.angle)
+            res$PrimaryAxis$LabelsRotation <- as.numeric(args$categories.tick.angle)
 
         # Export a native PowerPoint date axis when PrepareData captured the underlying dates (transformTable),
         # unless the user set the categories axis number type to "Category" (i.e. treat the date labels as
